+++ src/parser/satisfactory/types/property/special/ConveyorChainActorSpecialProperties.ts
@@ -1,8 +1,8 @@
 import { ContextReader } from '../../../../context/context-reader';
 import { ContextWriter } from '../../../../context/context-writer';
+import { FConveyorBeltItem } from '../../structs/binary/FConveyorBeltItem';
+import { FConveyorChainSplineSegment } from '../../structs/binary/FConveyorChainSplineSegment';
 import { ObjectReference } from '../../structs/ObjectReference';
-import { vec3 } from '../../structs/vec3';
-
 
 
 export const isConveyorChainActorSpecialProperties = (obj: any): obj is ConveyorChainActorSpecialProperties => obj.type === 'ConveyorChainActorSpecialProperties';
@@ -11,12 +11,12 @@ export type ConveyorChainActorSpecialProperties = {
     type: 'ConveyorChainActorSpecialProperties';
     firstBelt: ObjectReference;
     lastBelt: ObjectReference;
-    beltsInChain: ConveyorChainSegmentSpecialProperties[];
+    beltsInChain: FConveyorChainSplineSegment[];
     totalLength: number;
     totalNumberItemsMaybe: number;
     firstChainItemIndex: number;
     lastChainItemIndex: number;
-    items: ConveyorItemSpecialProperties[];
+    items: FConveyorBeltItem[];
 };
 
 export namespace ConveyorChainActorSpecialProperties {
@@ -26,40 +26,9 @@ export namespace ConveyorChainActorSpecialProperties {
         const firstBelt = ObjectReference.read(reader);
         const countBeltsInChain = reader.readInt32();
 
-        const beltsInChain: ConveyorChainSegmentSpecialProperties[] = [];
+        const beltsInChain: FConveyorChainSplineSegment[] = [];
         for (let i = 0; i < countBeltsInChain; i++) {
-            const chainActorRef = ObjectReference.read(reader);
-            const beltRef = ObjectReference.read(reader);
-            const splinePointsCount = reader.readInt32();
-
-            const splinePoints: { location: vec3; arriveTangent: vec3; leaveTangent: vec3; }[] = [];
-            for (let j = 0; j < splinePointsCount; j++) {
-                splinePoints.push({
-                    location: vec3.Parse(reader),
-                    arriveTangent: vec3.Parse(reader),
-                    leaveTangent: vec3.Parse(reader),
-                });
-            }
-
-            // indices which items of this chain are on this belt.
-            const offsetAtStart = reader.readFloat32();
-            const startsAtLength = reader.readFloat32();
-            const endsAtLength = reader.readFloat32();
-            const firstItemIndex = reader.readInt32();
-            const lastItemIndex = reader.readInt32();
-            const beltIndexInChain = reader.readInt32();
-
-            beltsInChain.push({
-                chainActorRef,
-                beltRef,
-                splinePoints,
-                offsetAtStart,
-                startsAtLength,
-                endsAtLength,
-                firstItemIndex,
-                lastItemIndex,
-                beltIndexInChain
-            });
+            beltsInChain.push(FConveyorChainSplineSegment.read(reader));
         }
 
         const totalLength = reader.readFloat32();
@@ -69,12 +38,9 @@ export namespace ConveyorChainActorSpecialProperties {
         const lastChainItemIndex = reader.readInt32();
         const countItemsInChain = reader.readInt32();
 
-        const items: ConveyorItemSpecialProperties[] = [];
+        const items: FConveyorBeltItem[] = [];
         for (let n = 0; n < countItemsInChain; n++) {
-            const item = ObjectReference.read(reader);
-            reader.readInt32(); //0
-            const position = reader.readInt32();
-            items.push({ itemReference: item, position });
+            items.push(FConveyorBeltItem.read(reader));
         }
 
         return {
@@ -97,23 +63,7 @@ export namespace ConveyorChainActorSpecialProperties {
         writer.writeInt32(property.beltsInChain.length);
 
         for (const belt of property.beltsInChain) {
-            ObjectReference.write(writer, belt.chainActorRef);
-            ObjectReference.write(writer, belt.beltRef);
-            writer.writeInt32(belt.splinePoints.length);
-
-
-            for (const splinepoint of belt.splinePoints) {
-                vec3.Serialize(writer, splinepoint.location);
-                vec3.Serialize(writer, splinepoint.arriveTangent);
-                vec3.Serialize(writer, splinepoint.leaveTangent);
-            }
-
-            writer.writeFloat32(belt.offsetAtStart);
-            writer.writeFloat32(belt.startsAtLength);
-            writer.writeFloat32(belt.endsAtLength);
-            writer.writeInt32(belt.firstItemIndex);
-            writer.writeInt32(belt.lastItemIndex);
-            writer.writeInt32(belt.beltIndexInChain);
+            FConveyorChainSplineSegment.write(writer, belt);
         }
 
         writer.writeFloat32(property.totalLength);
@@ -123,29 +73,7 @@ export namespace ConveyorChainActorSpecialProperties {
         writer.writeInt32(property.items.length);
 
         for (const item of property.items) {
-            ObjectReference.write(writer, item.itemReference);
-            writer.writeInt32(0);
-            writer.writeInt32(item.position);
+            FConveyorBeltItem.write(writer, item);
         }
     };
 }
-export type ConveyorChainSegmentSpecialProperties = {
-    chainActorRef: ObjectReference;
-    beltRef: ObjectReference;
-    splinePoints: {
-        location: vec3;
-        arriveTangent: vec3;
-        leaveTangent: vec3;
-    }[];
-    offsetAtStart: number;
-    startsAtLength: number;
-    endsAtLength: number;
-    firstItemIndex: number;
-    lastItemIndex: number;
-    beltIndexInChain: number;
-};
-export type ConveyorItemSpecialProperties = {
-    position: number;
-    itemReference: ObjectReference;
-};
-